Clean up

+++ ViewModels/MainViewModel.cs
@@ -1,7 +1,7 @@
-using System.Diagnostics;
 using System.IO;
 using System.IO.Compression;
 using System.Net.Http;
+using System.Reflection;
 using System.Windows.Input;
 using Microsoft.Office.Core;
 using Microsoft.Office.Interop.PowerPoint;
@@ -13,7 +13,6 @@ namespace SlidoCodingAssessment.ViewModels;
 
 public class MainViewModel : ViewModelBase
 {
-
     #region Commands
 
     private ICommand _downloadCommand;
@@ -26,14 +25,14 @@ public class MainViewModel : ViewModelBase
     public ICommand SaveLogsCommand =>
         _saveLogsCommand ??= new CommandHandler(GetLogs, true);
 
-    
     #endregion
-    
+
     #region Methods
+
     private async void DownloadFileAndOpenFile()
     {
         LoggerUtils.Instance.LogInfo(@"Downloading and opening presentation.");
-        
+
         var fileName = "How to use Slido.pptx";
         if (File.Exists(fileName)) File.Delete(fileName);
 
@@ -49,26 +48,26 @@ public class MainViewModel : ViewModelBase
                 LoggerUtils.Instance.LogError(@"Failed to download the file.", ex.Message);
             }
         }
-        
+
         LoggerUtils.Instance.LogInfo(@"File downloaded.");
         LoggerUtils.Instance.LogInfo(@"Opening presentation.");
 
-        string strExeFilePath = System.Reflection.Assembly.GetExecutingAssembly().Location;
-        string? strWorkPath = System.IO.Path.GetDirectoryName(strExeFilePath);
-        string fullFilePath = System.IO.Path.Combine(strWorkPath, fileName);
-        
+        var strExeFilePath = Assembly.GetExecutingAssembly().Location;
+        var strWorkPath = Path.GetDirectoryName(strExeFilePath);
+        var fullFilePath = Path.Combine(strWorkPath, fileName);
+
         try
         {
             var ppApp = new Application();
             ppApp.Visible = MsoTriState.msoTrue;
-            Presentations ppPresens = ppApp.Presentations;
-            Presentation objPres = ppPresens.Open(fullFilePath, MsoTriState.msoFalse, MsoTriState.msoTrue, MsoTriState.msoTrue);
+            var ppPresens = ppApp.Presentations;
+            var objPres = ppPresens.Open(fullFilePath, MsoTriState.msoFalse, MsoTriState.msoTrue);
         }
         catch (Exception ex)
         {
             LoggerUtils.Instance.LogError(@"Failed to open presentation.", ex.Message);
         }
-        
+
         LoggerUtils.Instance.LogInfo(@"Presentation opened.");
 
         //If I want to run slideShow
@@ -76,7 +75,7 @@ public class MainViewModel : ViewModelBase
         // Microsoft.Office.Interop.PowerPoint.SlideShowSettings objSSS;
         // objSSS = objPres.SlideShowSettings;
         // objSSS.Run();
-        
+
         //Other way to open PowerPoint but the exe path is not always in the same place
         // string powerPointPath = @"C:\Program Files\Microsoft Office\root\Office16\POWERPNT.EXE";
         // Process powerPoint = new Process();
@@ -88,43 +87,31 @@ public class MainViewModel : ViewModelBase
     private void GetLogs()
     {
         LoggerUtils.Instance.LogInfo(@"Creating logs.");
-        
+
         var keyPath = @"SOFTWARE\Microsoft\Office\PowerPoint\Addins";
         var regKey = Registry.CurrentUser.OpenSubKey(keyPath);
         var addins = regKey?.GetSubKeyNames();
 
         var fileName = @"log.txt";
-        string strExeFilePath = System.Reflection.Assembly.GetExecutingAssembly().Location;
-        string? strWorkPath = System.IO.Path.GetDirectoryName(strExeFilePath);
-        string fullFilePath = System.IO.Path.Combine(strWorkPath, fileName);
-        
+        var strExeFilePath = Assembly.GetExecutingAssembly().Location;
+        var strWorkPath = Path.GetDirectoryName(strExeFilePath);
+        var fullFilePath = Path.Combine(strWorkPath, fileName);
+
 
         var fileContent = LoggerUtils.Instance.GetLogContent().ToList();
-        if (addins != null)
-        {
-            fileContent.AddRange(addins);
-        }
+        if (addins != null) fileContent.AddRange(addins);
         File.AppendAllLines(fullFilePath, fileContent);
-        
-        var zipFullPath = System.IO.Path.Combine(strWorkPath, "Log.zip");
 
-        if (File.Exists(zipFullPath))
-        {
-            File.Delete(zipFullPath);
-        }
+        var zipFullPath = Path.Combine(strWorkPath, "Log.zip");
+
+        if (File.Exists(zipFullPath)) File.Delete(zipFullPath);
         using (var zip = ZipFile.Open(zipFullPath, ZipArchiveMode.Create))
         {
             zip.CreateEntryFromFile(fullFilePath, fileName);
         }
 
-        if (File.Exists(fullFilePath))
-        {
-            File.Delete(fullFilePath);
-        }
-        
+        if (File.Exists(fullFilePath)) File.Delete(fullFilePath);
     }
-    
+
     #endregion
-    
-    
 }
